Nav bar

+++ back-end/service/task.service.ts
@@ -1,5 +1,6 @@
 import projectDb from "../domain/data-access/project.db";
 import taskDb from "../domain/data-access/task.db";
+import userDb from "../domain/data-access/user.db";
 import { Task } from "../domain/model/task";
 import { TaskInput } from "../types";
 
@@ -59,4 +60,33 @@ const updateTask = async (task: TaskInput): Promise<Task> => {
     const ttask = await taskDb.updateTask(newTask);
     return ttask;
 }
-export default { getAllTasks, getTaskById, getTaskByProjectId, createTask, deleteById, updateTask };
+
+
+const getTaskByUserId = async (userId: number): Promise<Task[]> => {
+    const projects = [];
+    const tasks = [];
+    
+    const user = await userDb.getUserById(userId);
+
+
+    await Promise.all(user.teams.map(async (team) => {
+        const project = await projectDb.getProjectByTeamId(team.id);
+        projects.push(...project);
+    }));
+
+    // alternative way:
+    // for (const team of user.teams) {
+    //     const project = await projectDb.getProjectByTeamId(team.id);
+    //     projects.push(...project);
+    // }
+
+
+    await Promise.all(projects.map(async (project) => {
+        const task = await taskDb.getTaskByProject(project.id);
+        tasks.push(...task);
+    }));
+
+    return tasks;
+};
+
+export default { getAllTasks, getTaskById, getTaskByProjectId, createTask, deleteById, updateTask, getTaskByUserId };
